+++ src/test/java/org/seedstack/swagger/fixtures/MyFilter.java
@@ -6,7 +6,9 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.seedstack.swagger.internal;
+package org.seedstack.swagger.fixtures;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 import io.swagger.core.filter.SwaggerSpecFilter;
 import io.swagger.model.ApiDescription;
@@ -16,25 +18,35 @@ import io.swagger.models.parameters.Parameter;
 import io.swagger.models.properties.Property;
 import java.util.List;
 import java.util.Map;
-import org.seedstack.seed.Ignore;
+import javax.inject.Inject;
+import org.seedstack.seed.Application;
 
-@Ignore
 public class MyFilter implements SwaggerSpecFilter {
+    @Inject
+    private Application application;
+    public static boolean opAllowedCalled = false;
+    public static boolean paramAllowedCalled = false;
+
     @Override
     public boolean isOperationAllowed(Operation operation, ApiDescription api, Map<String, List<String>> params,
             Map<String, String> cookies, Map<String, List<String>> headers) {
-        return false;
+        assertThat(application).isNotNull();
+        opAllowedCalled = true;
+        return true;
     }
 
     @Override
     public boolean isParamAllowed(Parameter parameter, Operation operation, ApiDescription api,
             Map<String, List<String>> params, Map<String, String> cookies, Map<String, List<String>> headers) {
-        return false;
+        assertThat(application).isNotNull();
+        paramAllowedCalled = true;
+        return true;
     }
 
     @Override
     public boolean isPropertyAllowed(Model model, Property property, String propertyName,
             Map<String, List<String>> params, Map<String, String> cookies, Map<String, List<String>> headers) {
-        return false;
+        assertThat(application).isNotNull();
+        return true;
     }
 }